+++ camp2ascii/pipeline.py
@@ -209,13 +209,17 @@ def execute_config(cfg: Config) -> list[Path]:
     
     output_paths = []
     nbytes_proc_total = 0  # for progress bar tracking
-    for path in cfg.input_files:
+    for i, path in enumerate(cfg.input_files):
         df, header = process_file(path, cfg.stop_cond, pbar=cfg.pbar)
         if (cfg.timedate_filenames is not None and cfg.time_interval is None):
             out_path = Path(cfg.out_dir) / ("TOA5_" + path.stem + "_" + df.index.min().strftime(cfg.timedate_filenames) + path.suffix)
         else:
             out_path = Path(cfg.out_dir) / ("TOA5_" + path.name)
+        
+        if out_path.exists():
+            out_path = out_path.with_stem(out_path.stem + f"_{i}")
         output_paths.append(out_path)
+
         out_path = write_toa5_file(df, header, out_path, cfg.store_timestamp, cfg.store_record_numbers)
         if cfg.pbar is not None:
             nbytes_proc_total += out_path.stat().st_size